repomap/pyproject.toml
[tool.poetry]
name = "repomap"
version = "0.1.0"
description = "Repository mapping tool using tree-sitter"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
py-tree-sitter-languages = "^1.2.0"
tree-sitter = "^0.20.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
aider-chat = { path = "../aider", develop = true }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
repomap = "repomap.__main__:main"
